% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_confusion_matrix_heatmap.R
\name{generate_confusion_matrix_heatmap}
\alias{generate_confusion_matrix_heatmap}
\title{Create a heatmap for confusion matrix}
\usage{
generate_confusion_matrix_heatmap(conf_df, title = "Confusion Matrix Heatmap")
}
\arguments{
\item{conf_df}{A data frame containing the confusion matrix.}

\item{title}{A custom title for the heatmap plot.}
}
\value{
A ggplot heatmap object visualizing the confusion matrix.
}
\description{
Create a heatmap for confusion matrix
}
\examples{
conf_df <- data.frame(
Prediction = c("unacc", "acc", "good", "vgood"),
Reference = c("unacc", "acc", "good", "vgood"),
Freq = c(50, 10, 5, 35)
)
generate_confusion_matrix_heatmap(conf_df)

}
