% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature_plots.R
\name{generate_feature_barplots}
\alias{generate_feature_barplots}
\title{Generate a list of barplots for cateforical features against class}
\usage{
generate_feature_barplots(data, features)
}
\arguments{
\item{data}{A data frame containing the target class variable}

\item{features}{A list of categorical features}
}
\value{
A list of \code{ggplot} objects, each corresponding to one of the specified features.
}
\description{
This function  creates a eries of bar plots to visualize the relationship
between specified cateforical features and a target class variable in a given dataset.
Each plot displays the count of observations per feature category, grouped by the class.
}
\examples{
# example code
df <- data.frame(
    buying = c("vhigh", "vhigh", "vhigh", "vhigh", "vhigh", "low", "med", "low", "med"),
    maint = c("vhigh", "vhigh", "vhigh", "vhigh", "vhigh", "low", "low", "med", "med"),
    doors = c("2", "2", "2", "2", "2", "3", "3", "4", "5more"),
    persons = c("2", "2", "2", "2", "2", "4", "4", "more", "more"),
    lug_boot = c("small", "small", "small", "med", "med", "big", "big", "med", "small"),
    safety = c("low", "med", "high", "low", "med", "high", "low", "med", "high"),
    class = c("unacc", "unacc", "acc", "good", "vgood", "acc", "good", "vgood", "unacc")
)

category_features <- c("buying", "maint", "doors", "persons", "lug_boot", "safety")
plots <- generate_feature_barplots(data = df, features = category_features)

}
